refactor(panel): split shared UI responsibilities
