+++ panel/lib/app/presentation/shell/sidebar_state.dart
@@ -1,0 +1,13 @@
+part of "sidebar.dart";
+
+@riverpod
+class SidebarSize extends _$SidebarSize {
+  @override
+  double build() {
+    return kSidebarDefaultSize;
+  }
+
+  void size(double size) {
+    state = max(size, kSidebarMinSize);
+  }
+}
